rewrote redux middlewares to use TypeScript
